refactor(reasoning): Refactor node and DTOs to shared library
Migrate `ExecutionNode` interface, `ExecutionNodeType` enum, `ExecutionNodeConfig` interface,
`ExecutionNodeClass` schema, and `ExecutionNodeDto` from `api/src/modules/reasoning/node`
and `api/src/modules/reasoning/thinking` to `libs/caidense-reasoning/src/node`
and `libs/caidense-reasoning/src/node/dto` respectively.

This change centralizes the definition of execution nodes and their DTOs
within the shared `caidense-reasoning` library. This ensures consistent
data models and reduces duplication across the `api` and `worker` services.

Update all relevant imports and references in `api/src/modules/prompt/text`,
`api/src/modules/reasoning/node`, `api/src/modules/reasoning/thinking`,
and `worker/src` to point to the new shared library locations.

Also, `VariableDto` is moved from `api/src/modules/base/dto/variable.dto.ts`
to `libs/caidense-reasoning/src/common/dto/common.dto.ts` for better
organization and reusability.

The `ExecutionNode` interface was removed from `libs/caidense-reasoning/src/graph/graph.interface.ts`
as its definition is now solely in `libs/caidense-reasoning/src/node/node.interface.ts`.
The `ExecutionGraph` interface now references `ExecutionNodeDto`.
