+++ api/src/modules/reasoning/thinking/dto/thinking.dto.ts
@@ -1,9 +1,9 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { DocumentStatus } from '@/modules/base/base.interface';
-import { VariableDto } from '../../../base/dto/variable.dto';
+import { VariableDto } from '@caidense/reasoning/common/dto/common.dto';
 import { ReasoningThinkingDocument } from '../thinking.schemas';
-import { ReasoningNodeDto } from '@/modules/reasoning/node/dto/node.dto';
+import { ExecutionNodeDto } from '@caidense/reasoning/node/dto/node.dto';
 
 
 /**
@@ -102,11 +102,11 @@ export class ReasoningThinkingDto {
 
   @ApiProperty({
     description: 'An array of nodes within the thinking flow diagram.',
-    type: [ReasoningNodeDto],
+    type: [ExecutionNodeDto],
     isArray: true,
   })
-  @Type(() => ReasoningNodeDto)
-  nodes: ReasoningNodeDto[];
+  @Type(() => ExecutionNodeDto)
+  nodes: ExecutionNodeDto[];
 
   @ApiProperty({
     description: 'An array of edges connecting nodes within the thinking flow diagram.',
@@ -165,7 +165,7 @@ export class ReasoningThinkingDto {
     this.name = plainObject.name;
     this.description = plainObject.description;
 
-    this.nodes = plainObject.nodes ? plainObject.nodes.map(node => new ReasoningNodeDto(node)) : [];
+    this.nodes = plainObject.nodes ? plainObject.nodes.map(node => new ExecutionNodeDto(node)) : [];
     this.edges = plainObject.edges ? plainObject.edges.map(edge => new ReasoningThinkingEdgeDto(edge)) : [];
     this.inputs = plainObject.inputs ? plainObject.inputs.map(input => new VariableDto(input)) : [];
     this.outputs = plainObject.outputs ? plainObject.outputs.map(output => new VariableDto(output)) : [];
