+++ api/src/modules/reasoning/thinking/thinking.controller.ts
@@ -1,7 +1,7 @@
 import { BaseController } from '@/modules/base/base.controller';
 import { CreateVariableDto } from '@/modules/base/dto/create-variable.dto';
 import { UpdateVariableDto } from '@/modules/base/dto/update-variable.dto';
-import { VariableDto } from '@/modules/base/dto/variable.dto';
+import { VariableDto } from '@caidense/reasoning/common/dto/common.dto';
 import { Body, Controller, Delete, Get, Param, Post, Put, UsePipes, ValidationPipe } from '@nestjs/common';
 import {
   ApiBody,
