+++ api/src/modules/reasoning/thinking/thinking.schemas.ts
@@ -1,101 +1,14 @@
 import { DocumentStatus } from '@/modules/base/base.interface';
 import { Variable } from '@caidense/reasoning/common/common.interface';
 import { VariableSchema } from '@caidense/reasoning/common/common.schemas';
-import { ExecutionEdge, ExecutionNode } from '@caidense/reasoning/graph/graph.interface';
+import { ExecutionEdge } from '@caidense/reasoning/graph/graph.interface';
+import { ExecutionEdgeSchema } from '@caidense/reasoning/graph/graph.schemas';
+import { ExecutionNodeSchema } from '@caidense/reasoning/node/node.schemas';
+import { ExecutionNode } from '@caidense/reasoning/node/node.interface';
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document, Schema as MongooseSchema, Types } from 'mongoose';
-import { ReasoningNodeConfig } from '../node/node.interface';
+import { Document, Types } from 'mongoose';
 import { ReasoningThinking } from './thinking.interface';
 
-
-@Schema({
-  timestamps: true,
-  toJSON: {
-    virtuals: true,
-  },
-})
-export class ThinkingNodeClass implements ExecutionNode {
-  /**
-   * The type of the node, determining its function and behavior.
-   * Maps to ThinkingNode.type.
-   */
-  @Prop({ type: String, required: true })
-  type: string;
-
-  /**
-   * An optional label displayed on the node.
-   * Maps to ThinkingNode.label.
-   */
-  @Prop({ type: String })
-  label?: string;
-
-  /**
-   * Optional configuration object specific to the node type.
-   * Stored as a Mixed type to allow flexible structures.
-   * Maps to ThinkingNode.config.
-   */
-  @Prop({ type: MongooseSchema.Types.Mixed })
-  config?: ReasoningNodeConfig;
-
-  /**
-   * Optional script code to be executed by this node type.
-   * Maps to script.
-   */
-  @Prop({ type: String })
-  script?: string;
-
-  /**
-   * Optional array defining the inputs specific to this node's logic.
-   * Stored as an array of embedded Variable documents.
-   * Maps to inputs.
-   */
-  @Prop({ type: [VariableSchema] })
-  inputs?: Variable[];
-
-  /**
-   * Optional array defining the outputs specific to this node's logic.
-   * Stored as an array of embedded Variable documents.
-   * Maps to outputs.
-   */
-  @Prop({ type: [VariableSchema] })
-  outputs?: Variable[];
-
-  @Prop(Date)
-  createdAt: Date;
-
-  @Prop(Date)
-  updatedAt: Date;
-}
-
-export const ThinkingNodeSchema = SchemaFactory.createForClass(ThinkingNodeClass);
-
-@Schema({
-  timestamps: true,
-  toJSON: {
-    virtuals: true,
-  },
-})
-export class ThinkingEdgeSchemaClass implements ExecutionEdge {
-  @Prop({ type: String, required: true })
-  source: string;
-
-  @Prop({ type: String })
-  sourceHandle?: string;
-
-  @Prop({ type: String, required: true })
-  target: string;
-
-  @Prop({ type: String })
-  targetHandle?: string;
-
-  @Prop({ type: String })
-  type?: string;
-
-  @Prop({ type: String })
-  label?: string;
-}
-
-export const ThinkingEdgeSchema = SchemaFactory.createForClass(ThinkingEdgeSchemaClass);
 
 @Schema({
   timestamps: true,
@@ -110,10 +23,10 @@ export class ReasoningThinkingDocument extends Document implements ReasoningThin
   @Prop({ type: String })
   description?: string;
 
-  @Prop({ type: [ThinkingNodeSchema] })
+  @Prop({ type: [ExecutionNodeSchema] })
   nodes: ExecutionNode[];
 
-  @Prop({ type: [ThinkingEdgeSchema] })
+  @Prop({ type: [ExecutionEdgeSchema] })
   edges: ExecutionEdge[];
 
   @Prop({ type: [VariableSchema] })
