+++ libs/caidense-reasoning/package.json
@@ -1,5 +1,10 @@
 {
   "dependencies": {
-    "@nestjs/mongoose": "^10.1.0"
+    "@nestjs/mongoose": "^10.1.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.2"
+  },
+  "devDependencies": {
+    "@nestjs/swagger": "^8.1.1"
   }
 }
