+++ libs/caidense-reasoning/src/execution/execution.interface.ts
@@ -5,15 +5,3 @@ export enum ExecutionStatus {
   CANCELLED = 'cancelled',
   SKIPPED = 'skipped',
 }
-
-export enum ExecutionNodeType {
-  SEQUENCE = 'sequence',
-  PARALLEL = 'parallel',
-  EXCLUSIVE = 'exclusive',
-  INCLUSIVE = 'inclusive',
-  START_EVENT = 'startEvent',
-  END_EVENT = 'endEvent',
-  TASK = 'task',
-  GATEWAY = 'gateway',
-  SCRIPT = 'script',
-}
