+++ libs/caidense-reasoning/src/node/dto/node.dto.ts
@@ -1,13 +1,13 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
-import { ReasoningNode } from '../node.interface';
-import { VariableDto } from '@caidense/reasoning/common/dto/variable.dto';
+import { VariableDto } from '@caidense/reasoning/common/dto/common.dto';
+
 
 /**
- * Defines the data structure for representing a Reasoning Node document in API responses.
- * This DTO is consistent with the ReasoningNode interface.
+ * Defines the data structure for representing a Execution Node document in API responses.
+ * This DTO is consistent with the ExecutionNode interface.
  */
-export class ReasoningNodeDto {
+export class ExecutionNodeDto {
   @ApiProperty({
     description: 'The unique identifier (ID) of the node.',
     example: '60c8d6c1f5a4c8a7f0b1c2d6', // Example of MongoDB ObjectId string
@@ -37,6 +37,15 @@ export class ReasoningNodeDto {
   config?: object;
 
   @ApiProperty({
+    description: 'Optional array of IDs representing the incoming sequence edges to this node.',
+    type: [String],
+    isArray: true,
+    required: false,
+  })
+  @Type(() => String)
+  incoming?: string[];
+
+  @ApiProperty({
     description: 'Optional array defining the inputs specific to this node\'s logic.',
     type: [VariableDto],
     isArray: true,
@@ -51,6 +60,15 @@ export class ReasoningNodeDto {
     example: 'let sum = a + b; return sum > 100;'
   })
   script: string;
+
+  @ApiProperty({
+    description: 'Optional array of IDs representing the outgoing sequence edges from this node.',
+    type: [String],
+    isArray: true,
+    required: false,
+  })
+  @Type(() => String)
+  outgoing?: string[];
 
   @ApiProperty({
     description: 'Optional array defining the outputs specific to this node\'s logic.',
@@ -69,15 +87,17 @@ export class ReasoningNodeDto {
   @Type(() => Date)
   updatedAt: Date;
 
-  constructor(document: ReasoningNode) {
+  constructor(document: any) {
     const plainObject = document.toJSON ? document.toJSON() : document;
 
     this._id = plainObject._id.toHexString();
     this.type = plainObject.type;
     this.label = plainObject.label;
     this.config = plainObject.config;
+    this.incoming = plainObject.incoming ? plainObject.incoming.map(incoming => incoming.toHexString()) : [];
     this.inputs = plainObject.inputs ? plainObject.inputs.map(input => new VariableDto(input)) : [];
     this.script = plainObject.script;
+    this.outgoing = plainObject.outgoing ? plainObject.outgoing.map(outgoing => outgoing.toHexString()) : [];
     this.outputs = plainObject.outputs ? plainObject.outputs.map(output => new VariableDto(output)) : [];
     this.createdAt = plainObject.createdAt;
     this.updatedAt = plainObject.updatedAt;
